Created preprocessor module, adapted in a0.py and a1xx.py and updated docs
